AI/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain 0.2 (호환 버전)
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.17
langgraph==0.2.28  # LangGraph for workflow orchestration

# OpenAI (호환 버전)
openai==1.40.0
httpx==0.27.0

# OpenSearch
opensearch-py==2.4.2

# Kafka (for enrichment consumer)
kafka-python==2.0.2
aiohttp==3.9.1

# Utils
python-dotenv==1.0.0
python-json-logger==2.0.7
tiktoken==0.7.0  # Token counting for history truncation
cachetools==5.3.3  # Embedding cache
jinja2==3.1.3  # HTML templating for document generation
markdown==3.7  # Markdown to HTML conversion for AI analysis results
